Rewrite offline manifest to work with client assets.
This requires the assets to be precompiled for production.
The rake task

  $ rake assets:precompile

will produce a set of files to public/assets.
In production environment, Rails asset helper tags will serve
a version with a digest in filename. The digest changes when
file contents change. Thus, the cache manifest is changed and
re-evaluated. In development only images are cached.